Update test requirements and build version
ILAMB has to be built against pymt==0.2 because it uses nonstandard
parameter types (lists, which aren't allowed in BMI). Also build
again netcdf-fortran=4.4.4=5. This may be temporary. Note that neither
of these dependencies affect the building of ILAMB, just the testing.

The bmi-ilamb version was updted to 0.1.2, which is reflected here.